RUM-12062: Reflection working and README.md done
