--- features/dd-sdk-android-rum-debug-widget/api/apiSurface
@@ -1,1 +1,0 @@
-fun com.datadog.android.rum.RumConfiguration.Builder.enableDebugWidget(android.app.Application, Boolean = BuildConfig.DEBUG)
