+++ features/dd-sdk-android-rum/api/dd-sdk-android-rum.api
@@ -125,6 +125,12 @@ public final class com/datadog/android/rum/RumConfiguration$Builder {
 	public final fun useViewTrackingStrategy (Lcom/datadog/android/rum/tracking/ViewTrackingStrategy;)Lcom/datadog/android/rum/RumConfiguration$Builder;
 }
 
+public final class com/datadog/android/rum/RumDebugWidgetKt {
+	public static final fun enableRumDebugWidget (Lcom/datadog/android/rum/RumConfiguration$Builder;Landroid/app/Application;Z)Lcom/datadog/android/rum/RumConfiguration$Builder;
+	public static synthetic fun enableRumDebugWidget$default (Lcom/datadog/android/rum/RumConfiguration$Builder;Landroid/app/Application;ZILjava/lang/Object;)Lcom/datadog/android/rum/RumConfiguration$Builder;
+	public static final fun isDebuggable (Landroid/app/Application;)Z
+}
+
 public final class com/datadog/android/rum/RumErrorSource : java/lang/Enum {
 	public static final field AGENT Lcom/datadog/android/rum/RumErrorSource;
 	public static final field CONSOLE Lcom/datadog/android/rum/RumErrorSource;
@@ -385,11 +391,6 @@ public abstract interface class com/datadog/android/rum/internal/instrumentation
 
 public abstract interface class com/datadog/android/rum/internal/instrumentation/insights/InsightsUpdatesListener {
 	public abstract fun onDataUpdated ()V
-}
-
-public final class com/datadog/android/rum/internal/instrumentation/insights/RumDebugWidgetKt {
-	public static final fun enableDebugWidget (Lcom/datadog/android/rum/RumConfiguration$Builder;Landroid/app/Application;Z)Lcom/datadog/android/rum/RumConfiguration$Builder;
-	public static synthetic fun enableDebugWidget$default (Lcom/datadog/android/rum/RumConfiguration$Builder;Landroid/app/Application;ZILjava/lang/Object;)Lcom/datadog/android/rum/RumConfiguration$Builder;
 }
 
 public abstract interface class com/datadog/android/rum/internal/monitor/AdvancedNetworkRumMonitor : com/datadog/android/rum/RumMonitor {
